Add files via upload
Problem Details 
The objective of this project is to set quizzes and allow users to take up the multiple-choice quiz using the command line.
Features:
Super-users should be able to set up the questions and correct answers for the question.
Super-users should be able to add the questions based on the topics and difficulty level.
CRUD options for topics should be provided to the super user.
The Quiz should support MCQ type questions with single correct options.
Users should be able to take up the quiz. 
The super user should be able to set up multiple quizzes.
The application should display the final score and the right answers.
When a user attempts a quiz, questions need to be fetched automatically from the one’s set by super-user.
Test taker’s details and the result should be displayed on the screen.